Feature: Add sensible default for first day of the using locale
This commits adds a default for the first day of the week setting.
This is better than arbitrarily choosing a first day of the week.
In addition, it reorders the days in the dailies screen with the
first day in the list being the day the user chose in the setting.

Issue: #546, Pull Request: #620