#1084 There are missing entries in language settings.
Settings-APP SETTINGS tab does not contain 'Korean' Settings tab. If the user wants to change to a different language and then set it back to Korean, there is no way to set it.

I read your comment well. Apart from modifications to the typo, it is not possible to change the language to Korean because the menu tab does not have the 'Korean' tab when using the existing Release version. Therefore, it would be better to add the 'Korean' tab to the menu tab.