Improve content descriptions of ReminderItemFormView
- Previously, the '+' button was focusable, even though it didn't do
anything, and Talkback read out 'Unlabelled button', regardless of
whether it was a '+' button or a 'x' button.

- Now, we only make it focusable when it becomes a delete button, and
Talkback reads out 'Delete Reminder, Button for 6.16pm' (or whatever
the time of the reminder it relates to is - this is achieved by setting
the android:labelFor property on the reminder TextView).