Added info about found quest items in a snackbar
When user finds an item as a result, the information is displayed in the same way,
as for other found items. If he/she finds both — quest item and some other treasure
both are displayed in the same dialog. Three subsequent notifications would be quite
unhandy and also it seemed easier to be accomplished this way.