Adapt to orientation changes in MainActivity.
MainActivity defines an alternate layout for screen more than 600dp wide, but was ignoring orientation changes.

This caused a bug where if the application was opened with a wide orientation, but then the orientation changed to a small one, the user would be presented with the wide layout compressed into an unusable screen.

fixes #2002